[UPD] Refactor redirect functions
