+++ src/pages/index.js
@@ -1,21 +1,6 @@
-import { useEffect } from "react";
-import { navigate } from "gatsby";
-
-const getRedirectLanguage = () => {
-  if (typeof navigator === `undefined`) {
-    return "en";
-  }
-
-  const lang = navigator && navigator.language && navigator.language.split("-")[0];
-  if (!lang) return "en";
-
-  switch (lang) {
-    case "fr":
-      return "fr";
-    default:
-      return "en";
-  }
-};
+import { useEffect } from 'react'
+import { navigate } from 'gatsby'
+import { getRedirectLanguage } from '../utils/helper'
 
 const IndexPage = () => {
   useEffect(() => {
