+++ src/utils/helper.js
@@ -1,0 +1,17 @@
+const getRedirectLanguage = () => {
+  if (typeof navigator === `undefined`) {
+    return "en";
+  }
+
+  const lang = navigator && navigator.language && navigator.language.split("-")[0];
+  if (!lang) return "en";
+
+  switch (lang) {
+    case "fr":
+      return "fr";
+    default:
+      return "en";
+  }
+};
+
+export { getRedirectLanguage }
